pyproject.toml
[project]
name = "django-editor-ymaps"
version = "3.0.0"
description = "Plugin for creating and editing Yandex maps with Django framework."
readme = "README.md"
authors = [{ name = "kebasyaty", email = "kebasyaty@gmail.com" }]
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.12,<4.0"
keywords = ['django', 'editor', 'yandex', 'maps', "map", 'admin', "panel"]
classifiers = [
    "Development Status :: 4 - Beta",
    # "Development Status :: 5 - Production/Stable",
    'Environment :: Web Environment',
    'Framework :: Django :: 6.0',
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: Implementation :: CPython",
    # https://pypi.org/classifiers/
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    'Topic :: Scientific/Engineering :: GIS',
    'Topic :: Internet :: WWW/HTTP :: Dynamic Content',
    "Typing :: Typed",
]
dependencies = [
    "django>=6.0.2",
    "django-admin-sortable>=2.3",
    "django-colorful>=1.4.0",
    "django-imagekit>=6.0.0",
    "django-ipware>=7.0.1",
    "lxml>=6.0.2",
    "pillow>=12.1.0",
    "python-slugify>=8.0.4",
    "xloft>=0.11.0",
]

[project.urls]
Homepage = "https://kebasyaty.github.io/django-editor-ymaps/"
Repository = "https://github.com/kebasyaty/django-editor-ymaps"
Source = "https://github.com/kebasyaty/django-editor-ymaps"
"Bug Tracker" = "https://github.com/kebasyaty/django-editor-ymaps/issues"
Changelog = "https://github.com/kebasyaty/django-editor-ymaps/blob/v3/CHANGELOG.md"

[tool.uv]
package = true
default-groups = ["dev", "lint", "test", "docs", "types"]
compile-bytecode = true
check-url = "https://pypi.org/simple"

[tool.uv.build-backend]
module-root = "app/djeym"

[tool.hatch.build.targets.wheel]
packages = ["app/djeym"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = []
lint = [
    "djlint>=1.36.4",
    "pre-commit>=4.5.1",
    "pyrefly>=0.51.1",
    "ruff>=0.15.0",
]
test = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-httpx>=0.36.0",
    "pytest-mock>=3.15.1",
]
docs = [
    "mike>=2.1.3",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "mkdocstrings[python]>=1.0.2",
    "pygments>=2.19.2",
    "pymdown-extensions>=10.20.1",
]
types = []

[tool.ruff]
fix = true
line-length = 120
indent-width = 4
src = ["app", "tests"]
include = ["app/**/*.py", "tests/**/*.py"]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    ".github",
    "assets",
    "docs",
    "site",
    "examples",
]

[tool.ruff.lint]
preview = true
select = [
    "D",     # pydocstyle
    "I",     # isort
    "E",     # flake8-errors
    "W",     # flake8-warnings
    "F",     # flake8-pyflakes
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "S",     # flake8-bandit
    "BLE",   # flake8-blind-except
    "T20",   # flake8-print
    "T10",   # flake8-debugger
    "PT",    # flake8-pytest-style
    "RUF",   # Ruff-specific
    "DTZ",   # flake8-datetimez
    "ERA",   # eradicate
    "FURB",  # refurb
    "ARG",   # flake8-unused-arguments
    "EM",    # flake8-errmsg
    "ICN",   # flake8-import-conventions
    "G",     # flake8-logging-format
    "PGH",   # pygrep-hooks
    "PIE",   # flake8-pie
    "PL",    # pylint
    "PTH",   # flake8-use-pathlib
    "RET",   # flake8-return
    "SIM",   # flake8-simplify
    "UP",    # pyupgrade
    "YTT",   # flake8-2020
    "EXE",   # flake8-executable
    "ASYNC", # flake8-async
]
ignore = [
    "D417",   # Missing argument descriptions in the docstring
    "D206",   # Docstring should be indented with spaces, not tabs
    "C408",   # Unnecessary `dict` call (rewrite as a literal)
    "RUF001", # ambiguous-unicode-character-docstring
    "RUF002", # ambiguous-unicode-character-docstring
    "RUF003", # ambiguous-unicode-character-docstring
    "RUF022", # unsorted-dunder-all
    "E722",   # Do not use bare except
    "S110",   # try-except-pass detected, consider logging the exception
    "S101",   # Use of `assert` detected
    "PLR",    # Design related pylint codes
    "EM101",  # Exception must not use a string literal, assign to variable first
    "EM102",  # Exception must not use an f-string literal, assign to variable first
    "ARG002", # Unused method argument
    "RUF012",
    "E501",
    "D101",
    "D103",
]
fixable = ["I"]
unfixable = [
    "RUF100", # Unused noqa
    "T20",    # Removes print statements
    "F841",   # Removes unused variables
]
flake8-unused-arguments.ignore-variadic-names = true
isort.required-imports = ["from __future__ import annotations"]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?)|dummy.*)$"

[tool.ruff.lint.flake8-bandit]
check-typed-exception = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
docstring-code-line-length = 120
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pyrefly]
project-includes = ["app", "tests"]
search-path = ["app", "tests"]
ignore-errors-in-generated-code = true
untyped_def_behavior = "check-and-infer-return-any"

[tool.pyrefly.errors]
missing-attribute = false

[[tool.pyrefly.sub-config]]
matches = "tests/**"

[tool.pyrefly.sub-config.errors]
missing-attribute = false

[tool.pytest.ini_options]
testpaths = ["tests"]
